+++ packages/studio/src/components/StudioRightPanel.tsx
@@ -8,6 +8,8 @@ import {
   STUDIO_INSPECTOR_PANELS_ENABLED,
   STUDIO_MOTION_PANEL_ENABLED,
 } from "./editor/manualEditingAvailability";
+import { useCallback } from "react";
+import { resolveDomEditSelection, type DomEditLayerItem } from "./editor/domEditing";
 import { useStudioContext } from "../contexts/StudioContext";
 import { usePanelLayoutContext } from "../contexts/PanelLayoutContext";
 import { useFileManagerContext } from "../contexts/FileManagerContext";
@@ -37,6 +39,7 @@ export function StudioRightPanel({
     captionEditMode,
     previewIframeRef,
     projectId,
+    activeCompPath,
     compositionDimensions,
     waitForPendingDomEditSaves,
     renderQueue,
@@ -59,9 +62,23 @@ export function StudioRightPanel({
     handleAskAgent,
     handleDomMotionCommit,
     handleDomMotionClear,
+    applyDomSelection,
   } = useDomEditContext();
 
   const { assets, fontAssets, handleImportFiles, handleImportFonts } = useFileManagerContext();
+
+  const isMasterView = !activeCompPath || activeCompPath === "index.html";
+  const handleSelectLayer = useCallback(
+    (layer: DomEditLayerItem) => {
+      const selection = resolveDomEditSelection(layer.element, {
+        activeCompositionPath: activeCompPath,
+        isMasterView,
+        preferClipAncestor: false,
+      });
+      if (selection) applyDomSelection(selection);
+    },
+    [activeCompPath, isMasterView, applyDomSelection],
+  );
 
   const renderJobs = renderQueue.jobs as RenderJob[];
 
@@ -147,6 +164,8 @@ export function StudioRightPanel({
                   onImportAssets={handleImportFiles}
                   fontAssets={fontAssets}
                   onImportFonts={handleImportFonts}
+                  activeCompositionPath={activeCompPath}
+                  onSelectLayer={handleSelectLayer}
                 />
               ) : motionPanelActive ? (
                 <MotionPanel
